Fixed a bug where the touch length was a u32 rather than s64

+++ source/spritetools_input.c
@@ -13,7 +13,7 @@
 static touchPosition INPUTTouchOrigin;
 static touchPosition INPUTTouchPosition;
 static touchPosition INPUTTouchLastPosition;
-static u32 INPUTTouchLength;
+static s64 INPUTTouchLength;
 static circlePosition INPUTCirclePosition;
 static accelVector INPUTAccelVector;
 static angularRate INPUTGyroRate;
